Addtocart action is created. order and orderıtem repository methods are created. Customer can add products to cart

+++ Business/Services/Abstract/ICustomerService.cs
@@ -10,5 +10,5 @@ public interface ICustomerService
     Task<IResult> CreateCustomerAsync(Guid userId, RegisterCustomerViewModel model);
     Task<IResult> DeleteCustomerAsync(Guid customerId);
     Task<IDataResult<PaginatedList<CustomerListViewModel>>> GetPaginatedCustomerListAsync(string? searchTerm, int page, int pageSize);
-    Task<int> CountAsync();
+    Task<decimal> CountAsync();
 }
