+++ DataAccess/Repositories/Abstract/ICustomerRepository.cs
@@ -6,7 +6,7 @@ namespace DataAccess.Repositories.Abstract;
 
 public interface ICustomerRepository : IBaseRepository<Customer>
 {
-    Task<int> CountAsync();
+    Task<decimal> CountAsync();
     Task<PaginatedList<CustomerListViewModel>> GetPaginatedCustomerDtosAsync(
    string? searchTerm, int page, int pageSize);
 }
